Further refactor of COG code
taking non-backend specific code to _shared.py
