+++ odc/geo/cog/__init__.py
@@ -1,8 +1,10 @@
 from ._rio import to_cog, write_cog, write_cog_layers
-from ._tifffile import CogMeta, save_cog_with_dask
+from ._shared import CogMeta, cog_gbox
+from ._tifffile import save_cog_with_dask
 
 __all__ = [
     "CogMeta",
+    "cog_gbox",
     "to_cog",
     "write_cog",
     "write_cog_layers",
